requirements.txt
streamlit>=1.32
anthropic>=0.40
pandas>=2.2
openpyxl>=3.1
python-dotenv>=1.0
pypdf>=4.0
pdfplumber>=0.11
Pillow>=10.0
